Break up utils.py
